26 separated c4g facade interface and implemented it for unity, enhanced UX, fixed window edge cases bugs
