+++ C4G.Unity/Assets/C4G/Editor/Scripts/Core.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 3c028c633dfa4aa689bb7ba559ff98d0
+timeCreated: 1742394615
