+++ C4G.Unity/Assets/C4G/Editor/Scripts/Core/IC4GFacade.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 7f19b5c4943c477594ef460e8604d92d
+timeCreated: 1742394847
