+++ C4G.Unity/Assets/C4G/Editor/Scripts/Core/ParsedPropertyInfo.cs
@@ -1,0 +1,14 @@
+namespace C4G.Editor.Core
+{
+    public sealed class ParsedPropertyInfo
+    {
+        public readonly string Name;
+        public readonly string Type;
+
+        public ParsedPropertyInfo(string name, string type)
+        {
+            Name = name;
+            Type = type;
+        }
+    }
+}
