+++ C4G.Unity/Assets/C4G/Editor/Scripts/Core/ParsedPropertyInfo.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: f8a46536003e45f384f33849ff0a15bd
+timeCreated: 1742395185
