+++ C4G.Unity/Assets/C4G/Editor/Scripts/Core/ParsedSheet.cs
@@ -1,0 +1,18 @@
+using System.Collections.Generic;
+
+namespace C4G.Editor.Core
+{
+    public sealed class ParsedSheet
+    {
+        public readonly string Name;
+        public readonly IReadOnlyCollection<ParsedPropertyInfo> Properties;
+        public readonly IReadOnlyCollection<IReadOnlyCollection<string>> Entities;
+
+        public ParsedSheet(string name, IReadOnlyCollection<ParsedPropertyInfo> properties, IReadOnlyCollection<IReadOnlyCollection<string>> entities)
+        {
+            Name = name;
+            Properties = properties;
+            Entities = entities;
+        }
+    }
+}
