--- C4G.Unity/Assets/C4G/Editor/Scripts/FolderReferenceDrawer.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 3976136d93854bd687d3c74f378e5c4d
-timeCreated: 1739967819
