+++ C4G.Unity/Assets/C4G/Editor/Scripts/Unity/C4GUnityFacade.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: cdc64a094b6946489688ca8be9505b4c
+timeCreated: 1742395474
