+++ C4G.Unity/Assets/C4G/Editor/Scripts/Unity/SheetParser.cs
@@ -1,34 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
+using C4G.Editor.Core;
 
-namespace C4G.Editor
+namespace C4G.Editor.Unity
 {
-    public sealed class ParsedPropertyInfo
-    {
-        public readonly string Name;
-        public readonly string Type;
-
-        public ParsedPropertyInfo(string name, string type)
-        {
-            Name = name;
-            Type = type;
-        }
-    }
-
-    public sealed class ParsedSheet
-    {
-        public readonly string Name;
-        public readonly IReadOnlyCollection<ParsedPropertyInfo> Properties;
-        public readonly IReadOnlyCollection<IReadOnlyCollection<string>> Entities;
-
-        public ParsedSheet(string name, IReadOnlyCollection<ParsedPropertyInfo> properties, IReadOnlyCollection<IReadOnlyCollection<string>> entities)
-        {
-            Name = name;
-            Properties = properties;
-            Entities = entities;
-        }
-    }
-
     public static class SheetParser
     {
         public static ParsedSheet ParseSheet(string sheetName, IList<IList<object>> sheetData)
