--- C4G.Unity/Assets/C4G/Runtime/Scripts/AssemblyInfo.cs
@@ -1,3 +1,0 @@
-using System.Runtime.CompilerServices;
-
-[assembly: InternalsVisibleTo("C4G.Editor")]
