--- C4G.Unity/Assets/C4G/Runtime/Scripts/AssemblyInfo.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 7a7468926d144ea393a7823828523aa1
-timeCreated: 1739969130
