+++ C4G.Unity/Assets/C4G/Runtime/Scripts/C4GSettings.cs
@@ -1,7 +1,7 @@
 using System.IO;
 using UnityEngine;
 
-namespace C4G.Runtime.Scripts
+namespace C4G.Runtime
 {
     [CreateAssetMenu(menuName = "C4G/Settings")]
     public class C4GSettings : ScriptableObject
@@ -11,11 +11,16 @@ namespace C4G.Runtime.Scripts
         [SerializeField] private string _clientSecret;
         [SerializeField, FolderReference] private string _generatedCodeFolderPath;
         [SerializeField, FolderReference] private string _generatedDataFolderPath;
-        
+
         public string TableId => _tableId;
         public string SheetName => _sheetName;
         public string ClientSecret => _clientSecret;
+
+        public bool IsGeneratedCodeFolderValid => !string.IsNullOrEmpty(_generatedCodeFolderPath) &&
+                                                  Directory.Exists(GeneratedCodeFolderFullPath);
         public string GeneratedCodeFolderFullPath => Path.GetFullPath(Path.Combine(Application.dataPath, _generatedCodeFolderPath));
+        public bool IsGeneratedDataFolderValid => !string.IsNullOrEmpty(_generatedCodeFolderPath) && 
+                                                  Directory.Exists(GeneratedDataFolderFullPath);
         public string GeneratedDataFolderFullPath => Path.GetFullPath(Path.Combine(Application.dataPath, _generatedDataFolderPath));
     }
 }
