+++ C4G.Unity/Assets/C4G/Tests/Editor/Unity/CodeGeneratorTests.cs
@@ -1,9 +1,10 @@
 using System;
 using System.Collections.Generic;
-using C4G.Editor;
+using C4G.Editor.Core;
+using C4G.Editor.Unity;
 using NUnit.Framework;
 
-namespace C4G.Tests.Editor
+namespace C4G.Tests.Editor.Unity
 {
     public class CodeGeneratorTests
     {
